network-automation-project/requirements.txt
netmiko>=4.3.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
